Add contract tests, a11y testing, middleware extraction, ADRs, and security hardening
- OpenAPI contract tests (8 tests, all passing) validate spec-to-code sync
- axe-core/Playwright accessibility tests for WCAG 2.1 AA compliance
- Coverage thresholds (40%/35%/35%/40%) in vitest.config.ts
- Global error boundary (global-error.tsx) for root layout crashes
- Middleware layer extracted to server/middleware/index.js with CSP nonce support
- Domain routes (999 lines) and system routes (492 lines) extracted from server.js
- SQLite blacklist hydration on cold start (survives restarts without Redis)
- 5 Architecture Decision Records documenting key design choices
- createRequire added for ESM/CJS route module interop
- server.js reduced from 35,539 to 33,936 lines (-1,603 lines)
