+++ concord-frontend/package.json
@@ -11,6 +11,7 @@
     "test:run": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:e2e": "playwright test",
+    "test:a11y": "npx playwright test tests/e2e/accessibility.spec.ts",
     "type-check": "tsc --noEmit",
     "validate-routes": "npx tsx scripts/validate-routes.ts",
     "score-lenses": "npx tsx scripts/score-lenses.ts",
@@ -18,80 +19,81 @@
     "validate-lens-quality:strict": "npx tsx scripts/validate-lens-quality.ts --strict"
   },
   "dependencies": {
-    "next": "^15.1.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "@tanstack/react-query": "^5.17.0",
-    "zustand": "^4.5.0",
-    "axios": "^1.6.5",
-    "socket.io-client": "^4.7.4",
-    "dexie": "^3.2.4",
-    "zod": "^3.22.4",
     "@conform-to/zod": "^1.0.0",
-    "cytoscape": "^3.28.1",
-    "three": "^0.160.0",
-    "@react-three/fiber": "^8.15.0",
-    "@react-three/drei": "^9.92.0",
-    "clsx": "^2.1.0",
-    "tailwind-merge": "^2.2.0",
-    "lucide-react": "^0.309.0",
-    "date-fns": "^3.2.0",
-    "framer-motion": "^11.0.3",
-    "@tiptap/react": "^2.2.0",
-    "@tiptap/starter-kit": "^2.2.0",
-    "@tiptap/extension-placeholder": "^2.2.0",
-    "@tiptap/extension-link": "^2.2.0",
-    "@tiptap/extension-image": "^2.2.0",
-    "@tiptap/extension-task-list": "^2.2.0",
-    "@tiptap/extension-task-item": "^2.2.0",
-    "@tiptap/extension-table": "^2.2.0",
-    "@tiptap/extension-table-row": "^2.2.0",
-    "@tiptap/extension-table-cell": "^2.2.0",
-    "@tiptap/extension-table-header": "^2.2.0",
-    "@tiptap/extension-mention": "^2.2.0",
-    "@tiptap/extension-highlight": "^2.2.0",
-    "@tiptap/extension-typography": "^2.2.0",
-    "@tiptap/suggestion": "^2.2.0",
-    "react-hotkeys-hook": "^4.4.4",
+    "@radix-ui/react-alert-dialog": "^1.0.5",
+    "@radix-ui/react-avatar": "^1.0.4",
+    "@radix-ui/react-context-menu": "^2.1.5",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-popover": "^1.0.7",
-    "@radix-ui/react-tooltip": "^1.0.7",
-    "@radix-ui/react-tabs": "^1.0.4",
-    "@radix-ui/react-switch": "^1.0.3",
-    "@radix-ui/react-slider": "^1.1.2",
     "@radix-ui/react-progress": "^1.0.3",
-    "@radix-ui/react-avatar": "^1.0.4",
-    "@radix-ui/react-context-menu": "^2.1.5",
-    "@radix-ui/react-alert-dialog": "^1.0.5",
-    "react-virtuoso": "^4.6.2",
-    "nanoid": "^5.0.4",
+    "@radix-ui/react-slider": "^1.1.2",
+    "@radix-ui/react-switch": "^1.0.3",
+    "@radix-ui/react-tabs": "^1.0.4",
+    "@radix-ui/react-tooltip": "^1.0.7",
+    "@react-three/drei": "^9.92.0",
+    "@react-three/fiber": "^8.15.0",
+    "@tanstack/react-query": "^5.17.0",
+    "@tiptap/extension-highlight": "^2.2.0",
+    "@tiptap/extension-image": "^2.2.0",
+    "@tiptap/extension-link": "^2.2.0",
+    "@tiptap/extension-mention": "^2.2.0",
+    "@tiptap/extension-placeholder": "^2.2.0",
+    "@tiptap/extension-table": "^2.2.0",
+    "@tiptap/extension-table-cell": "^2.2.0",
+    "@tiptap/extension-table-header": "^2.2.0",
+    "@tiptap/extension-table-row": "^2.2.0",
+    "@tiptap/extension-task-item": "^2.2.0",
+    "@tiptap/extension-task-list": "^2.2.0",
+    "@tiptap/extension-typography": "^2.2.0",
+    "@tiptap/react": "^2.2.0",
+    "@tiptap/starter-kit": "^2.2.0",
+    "@tiptap/suggestion": "^2.2.0",
+    "axios": "^1.6.5",
+    "clsx": "^2.1.0",
+    "cytoscape": "^3.28.1",
+    "date-fns": "^3.2.0",
+    "dexie": "^3.2.4",
+    "framer-motion": "^11.0.3",
     "fuse.js": "^7.0.0",
-    "react-confetti": "^6.1.0"
+    "lucide-react": "^0.309.0",
+    "nanoid": "^5.0.4",
+    "next": "^15.1.0",
+    "react": "^18.3.1",
+    "react-confetti": "^6.1.0",
+    "react-dom": "^18.3.1",
+    "react-hotkeys-hook": "^4.4.4",
+    "react-virtuoso": "^4.6.2",
+    "socket.io-client": "^4.7.4",
+    "tailwind-merge": "^2.2.0",
+    "three": "^0.160.0",
+    "zod": "^3.22.4",
+    "zustand": "^4.5.0"
   },
   "devDependencies": {
-    "typescript": "^5.3.3",
+    "@axe-core/playwright": "^4.11.1",
+    "@eslint/eslintrc": "^3.0.0",
+    "@eslint/js": "^9.0.0",
+    "@playwright/test": "^1.41.0",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.0",
+    "@testing-library/react": "^16.0.0",
     "@types/node": "^20.11.0",
     "@types/react": "^18.2.48",
     "@types/react-dom": "^18.2.18",
     "@types/three": "^0.160.0",
-    "tailwindcss": "^3.4.1",
-    "postcss": "^8.4.33",
+    "@vitejs/plugin-react": "^4.3.0",
+    "@vitest/coverage-v8": "^3.0.0",
     "autoprefixer": "^10.4.17",
     "eslint": "^9.0.0",
     "eslint-config-next": "^15.1.0",
-    "@eslint/js": "^9.0.0",
-    "vitest": "^3.0.0",
-    "@vitest/coverage-v8": "^3.0.0",
-    "@vitejs/plugin-react": "^4.3.0",
-    "@testing-library/react": "^16.0.0",
-    "@testing-library/dom": "^10.4.0",
-    "@testing-library/jest-dom": "^6.6.0",
-    "jsdom": "^25.0.0",
-    "@playwright/test": "^1.41.0",
-    "prettier": "^3.2.4",
     "globals": "^15.0.0",
+    "jsdom": "^25.0.0",
+    "postcss": "^8.4.33",
+    "prettier": "^3.2.4",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.3.3",
     "typescript-eslint": "^8.0.0",
-    "@eslint/eslintrc": "^3.0.0"
+    "vitest": "^3.0.0"
   }
 }
