+++ docs/adr/README.md
@@ -1,0 +1,13 @@
+# Architecture Decision Records
+
+This directory contains Architecture Decision Records (ADRs) for the Concord Cognitive Engine.
+
+## Index
+
+| ADR | Title | Status | Date |
+|-----|-------|--------|------|
+| [001](001-intentional-monolith.md) | Intentional Monolith Architecture | Accepted | 2026-01-05 |
+| [002](002-macro-first-execution.md) | Macro-First Execution Model | Accepted | 2026-01-05 |
+| [003](003-local-first-sovereignty.md) | Local-First Sovereignty | Accepted | 2026-01-05 |
+| [004](004-dtu-as-atomic-unit.md) | DTU as Atomic Knowledge Unit | Accepted | 2026-01-05 |
+| [005](005-redis-optional-infrastructure.md) | Redis as Optional Infrastructure | Accepted | 2026-02-17 |
