+++ server/package.json
@@ -7,6 +7,7 @@
     "start": "node server.js",
     "dev": "node --watch server.js",
     "test": "node --test 'tests/**/*.test.js' 'tests/**/*-tests.js'",
+    "test:coverage": "node --test --experimental-test-coverage tests/",
     "test:e2e": "node --test 'tests/e2e/**/*.test.js'",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
